+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Next release
+
+- Adds `retrieveEstimatedDeliveryDate` in Shipment service
+
 ## v6.3.0 (2023-04-24)
 
 - Fixes the Beta client, previously the beta functions were not accessible due to how they were exported. This has now changed so each service is prepended with `Beta` (eg: `BetaCarrierMetadata`)